Add EventHandler trait for streaming timeline
EventHandler trait reduces code of applications that depends on Mastors but current implementation is loosely and roughly.
Mastors also continues to provide usage without the EventHandler.
