+++ src/syncronous/methods/api/v1/streaming.rs
@@ -7,7 +7,7 @@ use crate::{
     Connection,
     Error,
     Result,
-    streaming_timeline::*,
+    streaming::*,
     utils,
 };
